requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
#asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
#backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bleach==6.1.0
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.15
click==8.1.7
coloredlogs==15.0.1
#comm @ file:///croot/comm_1671231121260/work
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.3
#debugpy @ file:///croot/debugpy_1690905042057/work
#decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.4.2
entrypoints==0.4
#exceptiongroup @ file:///croot/exceptiongroup_1668714342571/work
#executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastapi==0.109.0
fastjsonschema==2.19.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
google-auth==2.26.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.2
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
#ipykernel @ file:///croot/ipykernel_1691121631942/work
#ipython @ file:///croot/ipython_1694181358621/work
ipython-genutils==0.2.0
#ipywidgets @ file:///croot/ipywidgets_1679394798311/work
isoduration==20.11.0
#jedi @ file:///tmp/build/80754af9/jedi_1644297102865/work
Jinja2==3.1.3
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter_client==7.4.9
#jupyter_core @ file:///croot/jupyter_core_1698937308754/work
jupyter_server==2.12.4
jupyter_server_terminals==0.5.1
jupyterlab==3.5.0
#jupyterlab-widgets @ file:///croot/jupyterlab_widgets_1700168618520/work
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.0
langchain-community==0.0.11
langchain-core==0.1.10
langchain-openai==0.0.2.post1
langsmith==0.0.80
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.2
#matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mistune==3.0.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.14.1
nbformat==5.9.2
#nest-asyncio @ file:///croot/nest-asyncio_1672387112409/work
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.7.2
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.4.0
#packaging @ file:///croot/packaging_1693575174725/work
pandas==2.1.4
pandocfilters==1.5.0
#parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
#pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
#pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
pillow==10.2.0
#platformdirs @ file:///croot/platformdirs_1692205439124/work
posthog==3.3.2
prometheus-client==0.19.0
#prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
protobuf==4.25.2
#psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
#ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pulsar-client==3.4.0
#pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
#Pygments @ file:///croot/pygments_1684279966437/work
pymongo==4.6.1
pyparsing==3.1.1
pypdf==3.17.4
PyPika==0.48.9
pyproject_hooks==1.0.0
#python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.16.2
rsa==4.9
Send2Trash==1.8.2
#six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
#stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.35.1
sympy==1.12
tenacity==8.2.3
terminado==0.18.0
tiktoken==0.5.2
tinycss2==1.2.1
tokenizers==0.15.0
tomli==2.0.1
#tornado @ file:///croot/tornado_1696936946304/work
tqdm==4.66.1
traitlets==5.14.1
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
#typing_extensions @ file:///croot/typing_extensions_1705005625920/work
tzdata==2023.4
uri-template==1.3.0
urllib3==2.1.0
uvicorn==0.26.0
#uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
#widgetsnbextension @ file:///croot/widgetsnbextension_1679313860248/work
wrapt==1.16.0
yarl==1.9.4
#zipp @ file:///croot/zipp_1704206909481/work
docx2txt
flask
